+++ src/commands.rs
@@ -1,0 +1,10 @@
+mod download_signatures;
+mod generate_sample_sheet;
+mod run;
+mod split_vcf;
+mod visualize;
+
+pub use self::{
+    download_signatures::download_signatures, generate_sample_sheet::generate_sample_sheet,
+    run::run, split_vcf::split_vcf, visualize::visualize,
+};